refactor: rename maths single answer related variables
